Moved MovieRatingFinder into app module to fix resource issue
