+++ Movies/moviedb/src/main/res/values/strings.xml
@@ -1,4 +1,3 @@
 <resources>
     <string name="app_name">moviedb</string>
-    <string name="no_rating">No Rating</string>
 </resources>
